Updated failind execution tests: semanticTests/structs.

+++ test/libsolidity/semanticTests/structs/calldata/dynamic_nested.sol
@@ -1,9 +1,9 @@
 pragma abicoder               v2;
 
 contract C {
-	struct S2 { uint256 b; }
-	struct S { uint256 a; S2[] children; }
-	function f(S calldata s) external pure returns (uint256, uint256, uint256, uint256) {
+	struct S2 { uint b; }
+	struct S { uint a; S2[] children; }
+	function f(S calldata s) external pure returns (uint, uint, uint, uint) {
 		return (s.children.length, s.a, s.children[0].b, s.children[1].b);
 	}
 }
@@ -11,4 +11,4 @@ contract C {
 // compileViaYul: also
 // compileToEwasm: also
 // ----
-// f((uint256,(uint256)[])): 32, 17, 64, 2, 23, 42 -> 2, 17, 23, 42
+// f((uint,(uint)[])): refargs { 17, 0x01, 0x02, 23, 42 } -> 2, 17, 23, 42
