+++ test/libsolidity/semanticTests/structs/global.sol
@@ -1,8 +1,8 @@
 pragma abicoder               v2;
 
-struct S { uint256 a; uint256 b; }
+struct S { uint a; uint b; }
 contract C {
-    function f(S calldata s) external pure returns (uint256, uint256) {
+    function f(S calldata s) external pure returns (uint, uint) {
         return (s.a, s.b);
     }
 }
@@ -10,4 +10,4 @@ contract C {
 // compileViaYul: also
 // compileToEwasm: also
 // ----
-// f((uint256,uint256)): 42, 23 -> 42, 23
+// f((uint,uint)): refargs { 42, 23 } -> 42, 23
